feat(proxy): migrate middleware to proxy convention and fix cache lock
  - Rename middleware.ts → proxy.ts and export proxy() per Next.js 16 convention
  - Eliminates deprecation warning on every request
  - Fix DistributedCache.getOrSet() busy-wait: replace fixed 400ms poll
    with exponential backoff (100ms base, 1600ms cap)
  - Fix unsafe lock release: use Lua CAS script with unique token so DEL
    only fires if the lock still belongs to this instance
