+++ proxy.test.ts
@@ -1,14 +1,12 @@
 import { beforeEach, afterEach, describe, expect, it, vi } from 'vitest';
 import { NextRequest, NextResponse } from 'next/server';
-import { middleware, config } from './middleware';
+import { proxy, config } from './proxy';
 import { rateLimit } from '@/lib/rate-limit';
 
 vi.mock('@/lib/rate-limit', () => ({
   rateLimit: vi.fn(),
 }));
 
-// Helper: build a rateLimit mock that returns success for the first call
-// and a given result for the second call (used when both limiters fire).
 function mockBothLimiters(
   refreshResult: Awaited<ReturnType<typeof rateLimit>>,
   generalSuccess = true
@@ -21,13 +19,12 @@ function mockBothLimiters(
   });
 }
 
-describe('middleware', () => {
+describe('proxy', () => {
   let originalEnv: string | undefined;
 
   beforeEach(() => {
     vi.clearAllMocks();
     originalEnv = process.env.TRUSTED_PROXIES;
-    // Set trusted proxies to make standard multi-hop tests pass as trusted proxy chains
     process.env.TRUSTED_PROXIES = '5.6.7.8, 9.10.11.12';
   });
 
@@ -46,7 +43,7 @@ describe('middleware', () => {
     const nextSpy = vi.spyOn(NextResponse, 'next');
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    await middleware(request);
+    await proxy(request);
 
     expect(nextSpy).toHaveBeenCalled();
   });
@@ -60,7 +57,7 @@ describe('middleware', () => {
     });
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    const response = await middleware(request);
+    const response = await proxy(request);
 
     expect(response.status).toBe(429);
   });
@@ -74,7 +71,7 @@ describe('middleware', () => {
     });
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    const response = await middleware(request);
+    const response = await proxy(request);
 
     await expect(response.json()).resolves.toEqual({
       error: 'Too many requests',
@@ -90,7 +87,7 @@ describe('middleware', () => {
     });
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    await middleware(request);
+    await proxy(request);
 
     expect(rateLimit).toHaveBeenCalledWith(expect.any(String), 60, 60000);
   });
@@ -104,7 +101,7 @@ describe('middleware', () => {
     });
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    const response = await middleware(request);
+    const response = await proxy(request);
 
     expect(response.headers.get('X-RateLimit-Limit')).toBe('60');
     expect(response.headers.get('X-RateLimit-Remaining')).toBe('59');
@@ -120,7 +117,7 @@ describe('middleware', () => {
     });
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    const response = await middleware(request);
+    const response = await proxy(request);
 
     expect(response.headers.get('X-RateLimit-Limit')).toBe('60');
     expect(response.headers.get('X-RateLimit-Remaining')).toBe('58');
@@ -136,7 +133,7 @@ describe('middleware', () => {
     });
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-    const response = await middleware(request);
+    const response = await proxy(request);
 
     expect(response.headers.get('Content-Type')).toBe('application/json');
     expect(response.headers.get('X-RateLimit-Limit')).toBe('60');
@@ -158,7 +155,7 @@ describe('middleware', () => {
       },
     });
 
-    await middleware(request);
+    await proxy(request);
 
     expect(rateLimit).toHaveBeenCalledWith('1.2.3.4', 60, 60000);
   });
@@ -171,7 +168,6 @@ describe('middleware', () => {
       reset: 123456789,
     });
 
-    // Clear trusted proxies so 5.6.7.8 is untrusted
     process.env.TRUSTED_PROXIES = '';
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat', {
@@ -180,9 +176,8 @@ describe('middleware', () => {
       },
     });
 
-    await middleware(request);
+    await proxy(request);
 
-    // Should resolve to the untrusted proxy IP (5.6.7.8) instead of the spoofed client IP (1.2.3.4)
     expect(rateLimit).toHaveBeenCalledWith('5.6.7.8', 60, 60000);
   });
 
@@ -200,7 +195,7 @@ describe('middleware', () => {
       },
     });
 
-    await middleware(request);
+    await proxy(request);
 
     expect(rateLimit).toHaveBeenCalledWith('9.9.9.9', 60, 60000);
   });
@@ -215,7 +210,7 @@ describe('middleware', () => {
 
     const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
 
-    await middleware(request);
+    await proxy(request);
 
     expect(rateLimit).toHaveBeenCalledWith('127.0.0.1', 60, 60000);
   });
@@ -235,7 +230,7 @@ describe('middleware', () => {
       },
     });
 
-    await middleware(request);
+    await proxy(request);
 
     expect(rateLimit).toHaveBeenCalledWith('1.2.3.4', 60, 60000);
   });
@@ -254,33 +249,28 @@ describe('middleware', () => {
       },
     });
 
-    await middleware(request);
+    await proxy(request);
 
     expect(rateLimit).toHaveBeenCalledWith('1.2.3.4', 60, 60000);
   });
 
-  it('includes compare API matcher in middleware config', () => {
+  it('includes compare API matcher in proxy config', () => {
     expect(config.matcher).toContain('/api/compare/:path*');
   });
 
-  it('includes wrapped and student API matchers in middleware config', () => {
+  it('includes wrapped and student API matchers in proxy config', () => {
     expect(config.matcher).toContain('/api/wrapped/:path*');
     expect(config.matcher).toContain('/api/student/:path*');
   });
 
-  // ─── ?refresh=true rate-limit tests ───────────────────────────────────────
-
   describe('?refresh=true cache-bypass rate limit', () => {
     it('applies the refresh limiter (5 req/min) before the general limiter', async () => {
-      // Both limiters succeed
       mockBothLimiters({ success: true, limit: 5, remaining: 4, reset: 123456789 });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      await middleware(request);
+      await proxy(request);
 
-      // First call must be the refresh limiter with the prefixed key
       expect(rateLimit).toHaveBeenNthCalledWith(1, 'refresh:127.0.0.1', 5, 60000);
-      // Second call must be the general limiter with the bare IP
       expect(rateLimit).toHaveBeenNthCalledWith(2, '127.0.0.1', 60, 60000);
     });
 
@@ -288,7 +278,7 @@ describe('middleware', () => {
       mockBothLimiters({ success: false, limit: 5, remaining: 0, reset: 123456789 });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      const response = await middleware(request);
+      const response = await proxy(request);
 
       expect(response.status).toBe(429);
       const body = await response.json();
@@ -304,15 +294,13 @@ describe('middleware', () => {
       });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      const response = await middleware(request);
+      const response = await proxy(request);
 
-      // The refresh limiter has limit=5, so the header must reflect that — not 60
       expect(response.headers.get('X-RateLimit-Limit')).toBe('5');
       expect(response.status).toBe(429);
     });
 
     it('does NOT invoke the general limiter when the refresh limit is exceeded', async () => {
-      // Only the refresh limiter fires (returns fail); general limiter must not be called.
       vi.mocked(rateLimit).mockResolvedValueOnce({
         success: false,
         limit: 5,
@@ -321,9 +309,8 @@ describe('middleware', () => {
       });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      await middleware(request);
+      await proxy(request);
 
-      // The general-limiter call always uses (bare IP, 60, 60000) — it must not appear.
       expect(rateLimit).not.toHaveBeenCalledWith('127.0.0.1', 60, 60000);
     });
 
@@ -336,7 +323,7 @@ describe('middleware', () => {
       });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      const response = await middleware(request);
+      const response = await proxy(request);
 
       expect(response.headers.get('X-RateLimit-Limit')).toBe('5');
       expect(response.headers.get('X-RateLimit-Remaining')).toBe('0');
@@ -351,9 +338,8 @@ describe('middleware', () => {
       });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat');
-      await middleware(request);
+      await proxy(request);
 
-      // Only the general limiter should be called
       expect(rateLimit).toHaveBeenCalledTimes(1);
       expect(rateLimit).toHaveBeenCalledWith('127.0.0.1', 60, 60000);
     });
@@ -369,7 +355,7 @@ describe('middleware', () => {
       const request = new NextRequest(
         'http://localhost:3000/api/streak?user=octocat&refresh=false'
       );
-      await middleware(request);
+      await proxy(request);
 
       expect(rateLimit).toHaveBeenCalledTimes(1);
       expect(rateLimit).not.toHaveBeenCalledWith(expect.stringContaining('refresh:'), 5, 60000);
@@ -379,25 +365,22 @@ describe('middleware', () => {
       mockBothLimiters({ success: true, limit: 5, remaining: 3, reset: 123456789 });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      const response = await middleware(request);
+      const response = await proxy(request);
 
-      // Both limiters ran and overall request succeeded
       expect(rateLimit).toHaveBeenCalledTimes(2);
       expect(response.status).toBe(200);
     });
 
     it('returns 429 from general limiter even when refresh succeeds', async () => {
-      // refresh OK, general fails
       vi.mocked(rateLimit)
         .mockResolvedValueOnce({ success: true, limit: 5, remaining: 2, reset: 123456789 })
         .mockResolvedValueOnce({ success: false, limit: 60, remaining: 0, reset: 123456789 });
 
       const request = new NextRequest('http://localhost:3000/api/streak?user=octocat&refresh=true');
-      const response = await middleware(request);
+      const response = await proxy(request);
 
       expect(response.status).toBe(429);
       const body = await response.json();
-      // The general-limiter error body does NOT mention "refresh"
       expect(body.error).toBe('Too many requests');
     });
   });
